Split up the adapter class WeekAndSerializableConverter.java into WeekToSerializableAdapter.java and SerializableToWeekAdapter.java, to keep in line with the adapter design pattern and SRP.
