+++ src/main/java/com/csc207/api/WeekController.java
@@ -2,7 +2,7 @@ package com.csc207.api;
 
 import com.csc207.domain.TaskSerializable;
 import com.csc207.domain.Week;
-import com.csc207.domain.WeekAndSerializableConverter;
+import com.csc207.domain.WeekToSerializableAdapter;
 import com.csc207.domain.WeekSerializable;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RestController;
@@ -26,9 +26,9 @@ public class WeekController {
     @Transactional
     public void saveWeek(Week week) {
         // convert to week serializable
-        WeekSerializable convertedWeek = WeekAndSerializableConverter.WeekToWeekSerializable(week);
+        WeekSerializable convertedWeek = WeekToSerializableAdapter.WeekToWeekSerializable(week);
         // convert to task serializable
-        ArrayList<TaskSerializable> convertedTasks = WeekAndSerializableConverter.WeekToTaskSerializable(week);
+        ArrayList<TaskSerializable> convertedTasks = WeekToSerializableAdapter.WeekToTaskSerializable(week);
         // save weekSerializable
         this.weekSerializableInteractor.saveWeekSerializable(convertedWeek);
         // save taskSerializable
